+++ examples/enclosures.py
@@ -105,14 +105,6 @@ try:
 except HPOneViewException as e:
     print(e.msg)
 
-# Get the enclosure script
-print("Get the enclosure script")
-try:
-    script = enclosure.get_script()
-    pprint(script)
-except HPOneViewException as e:
-    print(e.msg)
-
 # Buid the SSO URL parameters
 print("Build the SSO (Single Sign-On) URL parameters for the enclosure")
 try: