+++ hpOneView/resources/servers/server_hardware.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -302,4 +302,23 @@ class ServerHardware(ResourcePatchMixin, ResourceUtilizationMixin, Resource):
             Resource
         """
         uri = "{}/physicalServerHardware".format(self.data["uri"])
+        return self._helper.do_get(uri)
+
+    @ensure_resource_client
+    def get_local_storage(self, ip=None):
+        """
+        Gets the updated version 2 local storage resource for the server, including storage controllers, 
+        drives, and volumes.
+
+        Args:
+            ip: IP address or host name of the server's iLO management processor
+
+        Returns:
+            Resource
+        """
+        uri = "{}/localStorageV2".format(self.data["uri"])
+
+        if ip:
+            uri = "{}?ip={}".format(uri, ip)
+
         return self._helper.do_get(uri)